pyproject.toml
[project]
name = "manus"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mediapipe==0.10.14",
    "opencv-python>=4.9",
    "scikit-learn>=1.4",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pyautogui>=0.9.54",
    "pynput>=1.7",
    "paho-mqtt>=2.0",
    "pydantic>=2.12.5",
    "pandas>=3.0.2",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "websockets>=16.0",
    "python-multipart>=0.0.26",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
